Added Result Type, Query, Mutation, Service

+++ refresh.bat
@@ -1,4 +1,3 @@
 @echo off
-del /f db.sqlite
-npx sequelize db:migrate
-npx sequelize db:seed:all
+if exist db.sqlite del /f db.sqlite
+npx sequelize db:migrate && npx sequelize db:seed:all
